Fix "412 precondition failed" when response has no ETag, but OC-ETag header
On some servers, the ETag header is stripped and/or modified by caches,
web servers, proxies, whatever. nextCloud sends an OC-ETag that we can
use instead, and should be using in the first place if it exists, as
we don't know if the ETag header has been changed. This is very
probably a fix for #257 and #135