Fix custom labels not being displayed
Custom labels are represented as two fields with a prefix, e.g., `group1.X-ABLabel` and `group1.TEL`. Previously, `canDisplay` would check whether it could display `group1.TEL`, which is never a known property type. Instead we now check the property type after the `.`, which in our example is `TEL` which is displayable.

Signed-off-by: Y0hy0h <y0hy0h@gmx.net>