fix: rendering of phone numbers
There was a bug in sabre/vobject that stopped us from storing URIs with phone numbers. So, we had to use a workaround and not include them.[^1]

The fix for sabre/vobject is merged and included in stable31 and stable30.[^2]

It seems like this problem started after updating to ical.js 2.2.0 because of fixes in how vcard3 design sets were detected.[^3][^4]

Now it seems to work, but I don't understand why removing the tel property from the vcard3 design set worked when detecting the design set wasn't working before.

[^1]: https://github.com/nextcloud/contacts/pull/1393#issuecomment-570945735
[^2]: https://github.com/sabre-io/vobject/pull/486
[^3]: https://github.com/kewisch/ical.js/pull/680
[^4]: https://github.com/kewisch/ical.js/issues/363

Signed-off-by: Daniel Kesselberg <mail@danielkesselberg.de>