fix: saving vcard fails due wrong rev value
This looks like another issue from the ical 2.2.0 update. An
ICAL.VCardTime object is mapped to the date-and-or-time type. Assigning
such an object to contact.rev triggers an error deep within ical.js when
trying to "undecorate" the object with the given vcard design set.[^1]

The current version would also work by changing `contact.rev = rev` to
`contact.rev = rev.toICALString()` because that would skip the
undecorating path.

As I understand it, date-and-or-time is available only in vCard 4.
Therefore, I used ICAL.Time, which should be available in vCard 3 as the
date-time type.

[^1]:https://github.com/kewisch/ical.js/blob/ac16646655277764a6a09ccfab88152ea3c62280/lib/ical/property.js#L208-L210

Signed-off-by: Daniel Kesselberg <mail@danielkesselberg.de>